requirements.txt
langchain==0.1.13
streamlit==1.32.2
pypdf==4.1.0
sentence_transformers
boto3
openai
python-dotenv
pandas
chromadb
uvicorn
xlrd
streamlit
streamlit_option_menu
pysqlite3-binary